Format bootstrap parameter table

Format bootstrap parameter estimate values and output selected columns to be shown in the bootstrap parameter table.

Left join this output bootstrap data.frame with the formatted output of non-bootstrap parameter estimates

format_boot_table(
  .boot_df,
  .select_cols = c("abb", "desc", "boot_value", "boot_ci"),
  .digit = getOption("pmparams.dig"),
  .maxex = getOption("pmparams.maxex")
)

Arguments

.boot_df

parameter estimates output from define_boot_table with modifications ready for formatting

.select_cols

columns to select for output. Default selects "abb", "desc", "boot_value", "boot_ci". To return all columns, specify "all" for .select_cols.

.digit

set significant digits for output (optional). Default is three digits

.maxex

set maxex for computation (optional). Default is NULL

Examples

#Using output from `define_boot_table` (defineBootOut),
boot_paramEst <- utils::read.csv(system.file("model/nonmem/boot/data/boot-106.csv",
                                 package = "pmparams"))
nonboot_paramEst <- utils::read.csv(system.file("model/nonmem/nonboot_param_est.csv",
                                    package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")

defineBootOut <- define_boot_table(.boot_estimates = boot_paramEst,
               .nonboot_estimates = nonboot_paramEst,
               .key = paramKey) %>% head()
#> [1] "Parameter table yaml path provided: /data/pmparams/inst/model/nonmem/pk-parameter-key-new.yaml"

format_boot_table(.boot_df = defineBootOut) %>% head()
#>           abb                                  desc boot_value      boot_ci
#> 1    KA (1/h)  First order absorption rate constant       1.57   1.39, 1.78
#> 2    V2/F (L)               Apparent central volume       61.5   58.3, 65.1
#> 3  CL/F (L/h)                    Apparent clearance       3.23   3.07, 3.42
#> 4    V3/F (L)            Apparent peripheral volume       67.3   65.0, 69.8
#> 5   Q/F (L/h) Apparent intercompartmental clearance       3.61   3.37, 3.86
#> 6 CL/F ~ eGFR                   eGFR effect on CL/F      0.484 0.408, 0.558

#To include all columns:

format_boot_table(.boot_df = defineBootOut, .select_cols="all")
#>   parameter_names  estimate      lower     upper   name         abb
#> 1          THETA1 0.4488035  1.3880675  1.782212 THETA1    KA (1/h)
#> 2          THETA2 4.1194900 58.2837748 65.053174 THETA2    V2/F (L)
#> 3          THETA3 1.1722050  3.0667366  3.416850 THETA3  CL/F (L/h)
#> 4          THETA4 4.2098350 64.9928978 69.826825 THETA4    V3/F (L)
#> 5          THETA5 1.2844250  3.3729214  3.864090 THETA5   Q/F (L/h)
#> 6          THETA6 0.4836870  0.4081938  0.558204 THETA6 CL/F ~ eGFR
#>                                    desc  panel    trans fixed     value
#> 1  First order absorption rate constant struct logTrans FALSE  1.566437
#> 2               Apparent central volume struct logTrans FALSE 61.527855
#> 3                    Apparent clearance struct logTrans FALSE  3.229105
#> 4            Apparent peripheral volume struct logTrans FALSE 67.345427
#> 5 Apparent intercompartmental clearance struct logTrans FALSE  3.612590
#> 6                   eGFR effect on CL/F    cov     none FALSE  0.483687
#>   transTHETA   TH    OM     S   LOG LOGIT lognormO   Osd logitOsd propErr
#> 1         NA TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 2         NA TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 3         NA TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 4         NA TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 5         NA TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 6         NA TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#>   addErr      boot_ci boot_value
#> 1  FALSE   1.39, 1.78       1.57
#> 2  FALSE   58.3, 65.1       61.5
#> 3  FALSE   3.07, 3.42       3.23
#> 4  FALSE   65.0, 69.8       67.3
#> 5  FALSE   3.37, 3.86       3.61
#> 6  FALSE 0.408, 0.558      0.484