Combine bootstrap estimates with non-bootstrap estimates and parameter key

Combines boot strap estimates and non boot strap estimates with information in parameter key. Performs some formatting of this combined data.frame. There are two main steps of this function:

Run bbr::param_estimates_compare to extract summary quantiles, the 5th, 50th, and 95th, of the bootstrap estimates for each model parameter.

Some parameter_names have punctuation such as OMEGA(1,1). A new column is added without punctuation, such as OMEGA11.

Following this, parameter details from the parameter key are joined to the boot strap parameter estimates. A dplyr::inner_join is used so that only parameters in the model output are kept in the table. This was done so that, if your base and final model used the same structural THETAs and random parameters, the same parameter key could be used for both.

This join adds the following columns: abb (abbreviation), desc (parameter description), panel, trans (transformation).

Reformat non-bootstrap estimates and left join onto combined bootstrap estimates and parameter key data.frame. Expected input is a data.frame with parameter estimates, with the columns: parameter_names, estimate.

With this information provided, a check is performed to determine whether parameters with special transformation rules were defined correctly. In addition, a series of TRUE/FALSE columns that will be used by subsequent functions.

define_boot_table(.boot_estimates, .nonboot_estimates, .key)

Arguments

.boot_estimates

parameter boot estimates- either path to file or data.frame

.nonboot_estimates

non-bootstrap final model - either path to file or bbr model_summary

.key

parameter key- either path to file or data.frame

See also

Examples

boot_paramEst <- utils::read.csv(system.file("model/nonmem/boot/data/boot-106.csv",
                                 package = "pmparams"))
nonboot_paramEst <- utils::read.csv(system.file("model/nonmem/nonboot_param_est.csv",
                                    package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")

define_boot_table(.boot_estimates = boot_paramEst,
               .nonboot_estimates = nonboot_paramEst,
               .key = paramKey)
#> [1] "Parameter table yaml path provided: /opt/rpkgs/4.1/pmparams/model/nonmem/pk-parameter-key-new.yaml"
#>    parameter_names    estimate       lower       upper    name          abb
#> 1           THETA1  0.44880350  1.38806751  1.78221206  THETA1     KA (1/h)
#> 2           THETA2  4.11949000 58.28377484 65.05317419  THETA2     V2/F (L)
#> 3           THETA3  1.17220500  3.06673660  3.41684975  THETA3   CL/F (L/h)
#> 4           THETA4  4.20983500 64.99289784 69.82682512  THETA4     V3/F (L)
#> 5           THETA5  1.28442500  3.37292139  3.86409016  THETA5    Q/F (L/h)
#> 6           THETA6  0.48368700  0.40819383  0.55820400  THETA6  CL/F ~ eGFR
#> 7           THETA7 -0.03861535 -0.16724268  0.08782413  THETA7   CL/F ~ Age
#> 8           THETA8  0.42044750  0.29365133  0.58652035  THETA8   CL/F ~ ALB
#> 9       SIGMA.1.1.  0.03996855  0.03760094  0.04238123 SIGMA11 Proportional
#> 10      OMEGA.1.1.  0.21791200  0.12965965  0.33063635 OMEGA11       IIV-KA
#> 11      OMEGA.2.1.  0.06564225  0.03282125  0.10749465 OMEGA21      V2/F-KA
#> 12      OMEGA.2.2.  0.08210580  0.06432011  0.10123122 OMEGA22     IIV-V2/F
#> 13      OMEGA.3.1.  0.12131700  0.08049071  0.17341253 OMEGA31      CL/F-KA
#> 14      OMEGA.3.2.  0.06963290  0.05252801  0.08823306 OMEGA32    CL/F-V2/F
#> 15      OMEGA.3.3.  0.11213300  0.08956925  0.14029418 OMEGA33     IIV-CL/F
#>                                     desc  panel       trans fixed       value
#> 1   First order absorption rate constant struct    logTrans FALSE  1.56643682
#> 2                Apparent central volume struct    logTrans FALSE 61.52785506
#> 3                     Apparent clearance struct    logTrans FALSE  3.22910497
#> 4             Apparent peripheral volume struct    logTrans FALSE 67.34542690
#> 5  Apparent intercompartmental clearance struct    logTrans FALSE  3.61259012
#> 6                    eGFR effect on CL/F    cov        none FALSE  0.48368700
#> 7                     Age effect on CL/F    cov        none FALSE -0.03861535
#> 8           Serum albumin effect on CL/F    cov        none FALSE  0.42044750
#> 9                               Variance     RV     propErr    NA  0.03996855
#> 10                Variance of absorption    IIV lognormalOm    NA  0.21791200
#> 11               Covariance of V2/F - KA    IIV        none    NA  0.06564225
#> 12            Variance of central volume    IIV lognormalOm    NA  0.08210580
#> 13               Covariance of CL/F - KA    IIV        none    NA  0.12131700
#> 14             Covariance of CL/F - V2/F    IIV        none    NA  0.06963290
#> 15                 Variance of clearance    IIV lognormalOm    NA  0.11213300
#>    transTHETA    TH    OM     S   LOG LOGIT lognormO   Osd logitOsd propErr
#> 1          NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 2          NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 3          NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 4          NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 5          NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE   FALSE
#> 6          NA  TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#> 7          NA  TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#> 8          NA  TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#> 9          NA FALSE FALSE  TRUE FALSE FALSE    FALSE FALSE    FALSE    TRUE
#> 10         NA FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE   FALSE
#> 11         NA FALSE  TRUE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#> 12         NA FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE   FALSE
#> 13         NA FALSE  TRUE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#> 14         NA FALSE  TRUE FALSE FALSE FALSE    FALSE FALSE    FALSE   FALSE
#> 15         NA FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE   FALSE
#>    addErr
#> 1   FALSE
#> 2   FALSE
#> 3   FALSE
#> 4   FALSE
#> 5   FALSE
#> 6   FALSE
#> 7   FALSE
#> 8   FALSE
#> 9   FALSE
#> 10  FALSE
#> 11  FALSE
#> 12  FALSE
#> 13  FALSE
#> 14  FALSE
#> 15  FALSE