Combine parameter estimates with parameter key

Combines model output parameter estimates with information in parameter key. Performs some formatting of this combined data.frame.

Expected input is a data.frame with parameter estimates, with the columns: parameter_names, estimate, stderr, random_effect_sd, random_effect_sdse, fixed, diag, shrinkage.

Some parameter_names have punctuation such as OMEGA(1,1). A new column is added without punctuation, such as OMEGA11.

Following this, parameter details from the parameter key are joined to the parameter estimates. A dplyr::inner_join is used so that only parameters in the model output are kept in the table. This was done so that, if your base and final model used the same structural THETAs and random parameters, the same parameter key could be used for both.

This join adds the following columns: abb (abbreviation), desc (parameter description), panel, trans (transformation).

With this information provided, a check is performed to determine whether parameters with special transformation rules were defined correctly. In addition, a series of TRUE/FALSE columns that will be used by subsequent functions.

define_param_table(.estimates, .key, .ci = 95, .zscore = NULL)

Arguments

.estimates

path to model directory, bbr NONMEM model, or data.frame of parameter estimates

.key

path to parameter key or data.frame of parameter key. Described in more detail in param_key

.ci

confidence interval. Default is 95

.zscore

z-score for the specified confidence interval. Only needed for confidence intervals that are NOT 90 or 95 percent

Examples

paramEst <- utils::read.csv(system.file("model/nonmem/param_est.csv", package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")

define_param_table(.estimates = paramEst, .key = paramKey, .ci = 95, .zscore = NULL) %>% head()
#> [1] "Parameter table yaml path provided: /data/pmparams/inst/model/nonmem/pk-parameter-key-new.yaml"
#>   X.1 X parameter_names estimate    stderr random_effect_sd random_effect_sdse
#> 1   1 1          THETA1 0.433592 0.0628744               NA                 NA
#> 2   2 2          THETA2 4.118760 0.0276273               NA                 NA
#> 3   3 3          THETA3 1.115030 0.0328475               NA                 NA
#> 4   4 4          THETA4 4.209910 0.0192485               NA                 NA
#> 5   5 5          THETA5 1.286310 0.0354262               NA                 NA
#> 6   6 6      OMEGA(1,1) 0.220606 0.0529799         0.469687          0.0563992
#>   fixed diag shrinkage    name        abb                                  desc
#> 1 FALSE   NA        NA  THETA1   KA (1/h)  First order absorption rate constant
#> 2 FALSE   NA        NA  THETA2   V2/F (L)               Apparent central volume
#> 3 FALSE   NA        NA  THETA3 CL/F (L/h)                    Apparent clearance
#> 4 FALSE   NA        NA  THETA4   V3/F (L)            Apparent peripheral volume
#> 5 FALSE   NA        NA  THETA5  Q/F (L/h) Apparent intercompartmental clearance
#> 6 FALSE TRUE   17.8988 OMEGA11     IIV-KA                Variance of absorption
#>    panel       trans transTHETA    TH    OM     S   LOG LOGIT lognormO   Osd
#> 1 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 2 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 3 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 4 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 5 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 6    IIV lognormalOm         NA FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE
#>   logitOsd propErr addErr    value        se corr_SD     lower     upper
#> 1    FALSE   FALSE  FALSE 0.433592 0.0628744       - 0.3103582 0.5568258
#> 2    FALSE   FALSE  FALSE 4.118760 0.0276273       - 4.0646105 4.1729095
#> 3    FALSE   FALSE  FALSE 1.115030 0.0328475       - 1.0506489 1.1794111
#> 4    FALSE   FALSE  FALSE 4.209910 0.0192485       - 4.1721829 4.2476371
#> 5    FALSE   FALSE  FALSE 1.286310 0.0354262       - 1.2168746 1.3557454
#> 6    FALSE   FALSE  FALSE 0.220606 0.0529799       - 0.1167654 0.3244466

#To choose a confidence interval that is not 95 or 90, look up z-score and add as function parameter
define_param_table(.estimates = paramEst, .key = paramKey, .ci = 82, .zscore = 0.915) %>% head()
#> [1] "Parameter table yaml path provided: /data/pmparams/inst/model/nonmem/pk-parameter-key-new.yaml"
#>   X.1 X parameter_names estimate    stderr random_effect_sd random_effect_sdse
#> 1   1 1          THETA1 0.433592 0.0628744               NA                 NA
#> 2   2 2          THETA2 4.118760 0.0276273               NA                 NA
#> 3   3 3          THETA3 1.115030 0.0328475               NA                 NA
#> 4   4 4          THETA4 4.209910 0.0192485               NA                 NA
#> 5   5 5          THETA5 1.286310 0.0354262               NA                 NA
#> 6   6 6      OMEGA(1,1) 0.220606 0.0529799         0.469687          0.0563992
#>   fixed diag shrinkage    name        abb                                  desc
#> 1 FALSE   NA        NA  THETA1   KA (1/h)  First order absorption rate constant
#> 2 FALSE   NA        NA  THETA2   V2/F (L)               Apparent central volume
#> 3 FALSE   NA        NA  THETA3 CL/F (L/h)                    Apparent clearance
#> 4 FALSE   NA        NA  THETA4   V3/F (L)            Apparent peripheral volume
#> 5 FALSE   NA        NA  THETA5  Q/F (L/h) Apparent intercompartmental clearance
#> 6 FALSE TRUE   17.8988 OMEGA11     IIV-KA                Variance of absorption
#>    panel       trans transTHETA    TH    OM     S   LOG LOGIT lognormO   Osd
#> 1 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 2 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 3 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 4 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 5 struct    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE
#> 6    IIV lognormalOm         NA FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE
#>   logitOsd propErr addErr    value        se corr_SD     lower     upper
#> 1    FALSE   FALSE  FALSE 0.433592 0.0628744       - 0.3760619 0.4911221
#> 2    FALSE   FALSE  FALSE 4.118760 0.0276273       - 4.0934810 4.1440390
#> 3    FALSE   FALSE  FALSE 1.115030 0.0328475       - 1.0849745 1.1450855
#> 4    FALSE   FALSE  FALSE 4.209910 0.0192485       - 4.1922976 4.2275224
#> 5    FALSE   FALSE  FALSE 1.286310 0.0354262       - 1.2538950 1.3187250
#> 6    FALSE   FALSE  FALSE 0.220606 0.0529799       - 0.1721294 0.2690826