Format bootstrap parameter table

Format bootstrap parameter estimate values and output selected columns to be shown in the bootstrap parameter table.

Left join this output bootstrap data.frame with the formatted output of non-bootstrap parameter estimates

format_boot_table(
  .boot_df,
  .select_cols = c("abb", "desc", "boot_value", "boot_ci_95"),
  .digit = getOption("pmparams.dig"),
  .maxex = getOption("pmparams.maxex")
)

Arguments

.boot_df

parameter estimates output from define_boot_table with modifications ready for formatting

.select_cols

columns to select for output. Default selects "abb", "desc", "boot_value", "boot_ci". To return all columns, specify "all" for .select_cols.

.digit

set significant digits for output (optional). Default is three digits

.maxex

set maxex for computation (optional). Default is NULL

Examples

#Using output from `define_boot_table` (defineBootOut),
boot_paramEst <- utils::read.csv(system.file("model/nonmem/boot/data/boot-106.csv",
                                 package = "pmparams"))
nonboot_paramEst <- utils::read.csv(system.file("model/nonmem/nonboot_param_est.csv",
                                    package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")

defineBootOut <- define_boot_table(.boot_estimates = boot_paramEst,
               .nonboot_estimates = nonboot_paramEst,
               .key = paramKey)
#> [1] "Parameter table yaml path provided: /data/pmparams/inst/model/nonmem/pk-parameter-key-new.yaml"

format_boot_table(.boot_df = defineBootOut)
#>             abb                                  desc boot_value     boot_ci_95
#> 1      KA (1/h)  First order absorption rate constant       1.57     1.39, 1.78
#> 2      V2/F (L)               Apparent central volume       61.5     58.3, 65.1
#> 3    CL/F (L/h)                    Apparent clearance       3.23     3.07, 3.42
#> 4      V3/F (L)            Apparent peripheral volume       67.3     65.0, 69.8
#> 5     Q/F (L/h) Apparent intercompartmental clearance       3.61     3.37, 3.86
#> 6   CL/F ~ eGFR                   eGFR effect on CL/F      0.484   0.408, 0.558
#> 7    CL/F ~ Age                    Age effect on CL/F    -0.0386 -0.167, 0.0878
#> 8    CL/F ~ ALB          Serum albumin effect on CL/F      0.420   0.294, 0.587
#> 9        IIV-KA                Variance of absorption      0.218           <NA>
#> 10     IIV-V2/F            Variance of central volume     0.0821           <NA>
#> 11     IIV-CL/F                 Variance of clearance      0.112           <NA>
#> 12      V2/F-KA               Covariance of V2/F - KA     0.0656           <NA>
#> 13      CL/F-KA               Covariance of CL/F - KA      0.121           <NA>
#> 14    CL/F-V2/F             Covariance of CL/F - V2/F     0.0696           <NA>
#> 15 Proportional                              Variance     0.0400           <NA>

#To include all columns:

format_boot_table(.boot_df = defineBootOut, .select_cols="all")
#>    parameter_names    estimate       lower       upper    name          abb
#> 1           THETA1  0.44880350  1.38806751  1.78221206  THETA1     KA (1/h)
#> 2           THETA2  4.11949000 58.28377484 65.05317419  THETA2     V2/F (L)
#> 3           THETA3  1.17220500  3.06673660  3.41684975  THETA3   CL/F (L/h)
#> 4           THETA4  4.20983500 64.99289784 69.82682512  THETA4     V3/F (L)
#> 5           THETA5  1.28442500  3.37292139  3.86409016  THETA5    Q/F (L/h)
#> 6           THETA6  0.48368700  0.40819383  0.55820400  THETA6  CL/F ~ eGFR
#> 7           THETA7 -0.03861535 -0.16724268  0.08782413  THETA7   CL/F ~ Age
#> 8           THETA8  0.42044750  0.29365133  0.58652035  THETA8   CL/F ~ ALB
#> 9       OMEGA.1.1.  0.21791200  0.12965965  0.33063635 OMEGA11       IIV-KA
#> 10      OMEGA.2.2.  0.08210580  0.06432011  0.10123122 OMEGA22     IIV-V2/F
#> 11      OMEGA.3.3.  0.11213300  0.08956925  0.14029418 OMEGA33     IIV-CL/F
#> 12      OMEGA.2.1.  0.06564225  0.03282125  0.10749465 OMEGA21      V2/F-KA
#> 13      OMEGA.3.1.  0.12131700  0.08049071  0.17341253 OMEGA31      CL/F-KA
#> 14      OMEGA.3.2.  0.06963290  0.05252801  0.08823306 OMEGA32    CL/F-V2/F
#> 15      SIGMA.1.1.  0.03996855  0.03760094  0.04238123 SIGMA11 Proportional
#>                                     desc  panel       trans fixed       value
#> 1   First order absorption rate constant struct    logTrans FALSE  1.56643682
#> 2                Apparent central volume struct    logTrans FALSE 61.52785506
#> 3                     Apparent clearance struct    logTrans FALSE  3.22910497
#> 4             Apparent peripheral volume struct    logTrans FALSE 67.34542690
#> 5  Apparent intercompartmental clearance struct    logTrans FALSE  3.61259012
#> 6                    eGFR effect on CL/F    cov        none FALSE  0.48368700
#> 7                     Age effect on CL/F    cov        none FALSE -0.03861535
#> 8           Serum albumin effect on CL/F    cov        none FALSE  0.42044750
#> 9                 Variance of absorption    IIV lognormalOm    NA  0.21791200
#> 10            Variance of central volume    IIV lognormalOm    NA  0.08210580
#> 11                 Variance of clearance    IIV lognormalOm    NA  0.11213300
#> 12               Covariance of V2/F - KA    IIV        none    NA  0.06564225
#> 13               Covariance of CL/F - KA    IIV        none    NA  0.12131700
#> 14             Covariance of CL/F - V2/F    IIV        none    NA  0.06963290
#> 15                              Variance     RV     propErr    NA  0.03996855
#>    transTHETA THETAERR    TH    OM     S   LOG LOGIT lognormO   Osd logitOsd
#> 1          NA    FALSE  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 2          NA    FALSE  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 3          NA    FALSE  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 4          NA    FALSE  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 5          NA    FALSE  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 6          NA    FALSE  TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE
#> 7          NA    FALSE  TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE
#> 8          NA    FALSE  TRUE FALSE FALSE FALSE FALSE    FALSE FALSE    FALSE
#> 9          NA    FALSE FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE
#> 10         NA    FALSE FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE
#> 11         NA    FALSE FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE
#> 12         NA    FALSE FALSE  TRUE FALSE FALSE FALSE    FALSE FALSE    FALSE
#> 13         NA    FALSE FALSE  TRUE FALSE FALSE FALSE    FALSE FALSE    FALSE
#> 14         NA    FALSE FALSE  TRUE FALSE FALSE FALSE    FALSE FALSE    FALSE
#> 15         NA    FALSE FALSE FALSE  TRUE FALSE FALSE    FALSE FALSE    FALSE
#>    propErr addErr addErrLogDV     boot_ci_95 boot_value
#> 1    FALSE  FALSE       FALSE     1.39, 1.78       1.57
#> 2    FALSE  FALSE       FALSE     58.3, 65.1       61.5
#> 3    FALSE  FALSE       FALSE     3.07, 3.42       3.23
#> 4    FALSE  FALSE       FALSE     65.0, 69.8       67.3
#> 5    FALSE  FALSE       FALSE     3.37, 3.86       3.61
#> 6    FALSE  FALSE       FALSE   0.408, 0.558      0.484
#> 7    FALSE  FALSE       FALSE -0.167, 0.0878    -0.0386
#> 8    FALSE  FALSE       FALSE   0.294, 0.587      0.420
#> 9    FALSE  FALSE       FALSE           <NA>      0.218
#> 10   FALSE  FALSE       FALSE           <NA>     0.0821
#> 11   FALSE  FALSE       FALSE           <NA>      0.112
#> 12   FALSE  FALSE       FALSE           <NA>     0.0656
#> 13   FALSE  FALSE       FALSE           <NA>      0.121
#> 14   FALSE  FALSE       FALSE           <NA>     0.0696
#> 15    TRUE  FALSE       FALSE           <NA>     0.0400