Format parameter table

Format parameter estimate values and output selected columns to be shown in the parameter table.

There are four main steps of this function:

When necessary, back transform parameters and their CIs, round parameters using pmtables::sig, and combine columns.

Format the THETA/OMEGA/SIGMA values to display as greek letters in Latex, with subscript numbers, and where necessary, the transformation applied to that parameter.

Determine which panel of the final table the parameter should be displayed in. This is informed by the panel argument you defined in your parameter key. Note that there are a finite number of options included by default (see below) but you can include additional panels as needed. Panel types include:

Residual variance

Interindividual covariance parameters

Interindividual variance parameters

Interoccasion variance parameters

Covariate effect parameters

Structural model parameters

Select columns for final tables.

format_param_table(
  .df,
  .select_cols = c("type", "abb", "greek", "desc", "value", "ci", "shrinkage"),
  .prse = FALSE,
  .digit = getOption("pmparams.dig"),
  .maxex = getOption("pmparams.maxex")
)

Arguments

.df

parameter estimates output from define_param_table with modifications ready for formatting

.select_cols

columns to select for output. Default selects "type", "abb", "greek", "desc", "value", "ci", "shrinkage". To return all columns, specify "all" for .select_cols

.prse

output pRSE. Default is FALSE

.digit

set significant digits for output (optional). Default is three digits

.maxex

set maxex for computation (optional). Default is NULL

Examples

#Using output from `define_param_table` (defineOut),
paramEst <- utils::read.csv(system.file("model/nonmem/param_est.csv", package = "pmparams"))
paramKey <-  system.file("model/nonmem/pk-parameter-key-new.yaml", package = "pmparams")
defineOut <- define_param_table(.estimates = paramEst, .key = paramKey, .ci = 95, .zscore = NULL)
#> [1] "Parameter table yaml path provided: /data/pmparams/inst/model/nonmem/pk-parameter-key-new.yaml"

format_param_table(.df = defineOut) %>% head()
#>                                  type        abb                greek
#> 1         Structural model parameters   KA (1/h) $\\exp(\\theta_{1})$
#> 2         Structural model parameters   V2/F (L) $\\exp(\\theta_{2})$
#> 3         Structural model parameters CL/F (L/h) $\\exp(\\theta_{3})$
#> 4         Structural model parameters   V3/F (L) $\\exp(\\theta_{4})$
#> 5         Structural model parameters  Q/F (L/h) $\\exp(\\theta_{5})$
#> 6 Interindividual variance parameters     IIV-KA    $\\Omega_{(1,1)}$
#>                                    desc              value           ci
#> 1  First order absorption rate constant               1.54   1.36, 1.75
#> 2               Apparent central volume               61.5   58.2, 64.9
#> 3                    Apparent clearance               3.05   2.86, 3.25
#> 4            Apparent peripheral volume               67.4   64.9, 69.9
#> 5 Apparent intercompartmental clearance               3.62   3.38, 3.88
#> 6                Variance of absorption 0.221 [CV\\%=49.7] 0.117, 0.324
#>   shrinkage
#> 1         -
#> 2         -
#> 3         -
#> 4         -
#> 5         -
#> 6      17.9

#To include all columns:

format_param_table(.df = defineOut, .select_cols="all") %>% head()
#>   X.1 X  num2 estimate    stderr random_effect_sd random_effect_sdse fixed diag
#> 1   1 1     1 0.433592 0.0628744               NA                 NA FALSE   NA
#> 2   2 2     2 4.118760 0.0276273               NA                 NA FALSE   NA
#> 3   3 3     3 1.115030 0.0328475               NA                 NA FALSE   NA
#> 4   4 4     4 4.209910 0.0192485               NA                 NA FALSE   NA
#> 5   5 5     5 1.286310 0.0354262               NA                 NA FALSE   NA
#> 6   6 6 (1,1) 0.220606 0.0529799         0.469687          0.0563992 FALSE TRUE
#>   shrinkage    name        abb                                  desc  panel
#> 1         -  THETA1   KA (1/h)  First order absorption rate constant struct
#> 2         -  THETA2   V2/F (L)               Apparent central volume struct
#> 3         -  THETA3 CL/F (L/h)                    Apparent clearance struct
#> 4         -  THETA4   V3/F (L)            Apparent peripheral volume struct
#> 5         -  THETA5  Q/F (L/h) Apparent intercompartmental clearance struct
#> 6      17.9 OMEGA11     IIV-KA                Variance of absorption    IIV
#>         trans transTHETA    TH    OM     S   LOG LOGIT lognormO   Osd logitOsd
#> 1    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 2    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 3    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 4    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 5    logTrans         NA  TRUE FALSE FALSE  TRUE FALSE    FALSE FALSE    FALSE
#> 6 lognormalOm         NA FALSE  TRUE FALSE FALSE FALSE     TRUE FALSE    FALSE
#>   propErr addErr              value        se corr_SD      lower      upper
#> 1   FALSE  FALSE               1.54 0.0628744       -  1.3639135  1.7451244
#> 2   FALSE  FALSE               61.5 0.0276273       - 58.2422183 64.9040164
#> 3   FALSE  FALSE               3.05 0.0328475       -  2.8595060  3.2524583
#> 4   FALSE  FALSE               67.4 0.0192485       - 64.8568764 69.9399530
#> 5   FALSE  FALSE               3.62 0.0354262       -  3.3766181  3.8796516
#> 6   FALSE  FALSE 0.221 [CV\\%=49.7] 0.0529799       -  0.1167654  0.3244466
#>     cv pRSE           ci sd  text                greek
#> 1    - 6.29   1.36, 1.75  - theta $\\exp(\\theta_{1})$
#> 2    - 2.76   58.2, 64.9  - theta $\\exp(\\theta_{2})$
#> 3    - 3.29   2.86, 3.25  - theta $\\exp(\\theta_{3})$
#> 4    - 1.93   64.9, 69.9  - theta $\\exp(\\theta_{4})$
#> 5    - 3.54   3.38, 3.88  - theta $\\exp(\\theta_{5})$
#> 6 49.7 24.0 0.117, 0.324  - Omega    $\\Omega_{(1,1)}$
#>                                  type type_f
#> 1         Structural model parameters      1
#> 2         Structural model parameters      1
#> 3         Structural model parameters      1
#> 4         Structural model parameters      1
#> 5         Structural model parameters      1
#> 6 Interindividual variance parameters      3